Starting to update tests + move other mock widgets
