--- packages/perseus/src/widgets/mock-widget/index.ts
@@ -1,1 +1,0 @@
-export {default} from "./mock-widget";
