+++ packages/perseus/src/widgets/mock-widgets/index.ts
@@ -1,0 +1,2 @@
+export {default as MockWidget} from "./mock-widget";
+export {default as MockAssetLoadingWidget} from "./mock-asset-loading-widget";
