+++ packages/perseus/src/widgets/mock-widgets/mock-asset-loading-widget.tsx
